requirements.txt
black
ipykernel
numpy
panda
pyarrow
matplotlib
plotly
seaborn
nbformat
python-dotenv
flake8
termcolor